// src/blaze-notes-app/src/app/api/models/to-do-task.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PriorityLevel } from '../models/priority-level';
export interface ToDoTask {
  completedAt?: string | null;
  createdAt?: Date | null;
  description?: string | null;
  dueDate?: string | null;
  id?: number;
  isCompleted?: boolean;
  priority?: PriorityLevel;
  title?: string | null;
}